+++ src/routes/stats.js
@@ -1,9 +1,17 @@
 'use strict';
+var _ = require('lodash');
 var auth = require('../services/auth');
 var logger = require('../services/logger');
 var error = require('../services/error');
 var path = require('../services/path');
 var StatSerializer = require('../serializers/stat');
+var Schemas = require('../generators/schemas');
+
+const CHART_TYPE_VALUE = 'Value';
+const CHART_TYPE_PIE = 'Pie';
+const CHART_TYPE_LINE = 'Line';
+const CHART_TYPE_LEADERBOARD = 'Leaderboard';
+const CHART_TYPE_OBJECTIVE = 'Objective';
 
 module.exports = function (app, model, Implementation, opts) {
   var modelName = Implementation.getModelName(model);
@@ -11,19 +19,28 @@ module.exports = function (app, model, Implementation, opts) {
   this.get = function (request, response, next) {
     var promise = null;
 
-    switch (request.body.type) {
-    case 'Value':
-      promise = new Implementation.ValueStatGetter(model, request.body, opts)
-        .perform();
-      break;
-    case 'Pie':
-      promise = new Implementation.PieStatGetter(model, request.body, opts)
-        .perform();
-      break;
-    case 'Line':
-      promise = new Implementation.LineStatGetter(model, request.body, opts)
-        .perform();
-      break;
+    function getAssociationModel(schema, associationName) {
+      const field = _.find(schema.fields, { field: associationName });
+      let relatedModelName;
+      if (field && field.reference) {
+        relatedModelName = field.reference.split('.')[0];
+      }
+
+      const models = Implementation.getModels();
+      return _.find(models, model => Implementation.getModelName(model) === relatedModelName);
+    }
+
+    let { type } = request.body;
+    if (type === CHART_TYPE_OBJECTIVE) { type = CHART_TYPE_VALUE; }
+
+    if (type === CHART_TYPE_LEADERBOARD) {
+      const schema = Schemas.schemas[model.name];
+      const modelRelationship = getAssociationModel(schema, request.body.relationship_field);
+
+      promise = new Implementation
+        .LeaderboardStatGetter(model, modelRelationship, request.body, opts).perform();
+    } else {
+      promise = new Implementation[`${type}StatGetter`](model, request.body, opts).perform();
     }
 
     if (!promise) {
@@ -32,6 +49,12 @@ module.exports = function (app, model, Implementation, opts) {
 
     promise
       .then(function (stat) {
+        if (request.body.type === CHART_TYPE_OBJECTIVE) {
+          stat.value.value = stat.value.countCurrent;
+          delete stat.value.countCurrent;
+          delete stat.value.countPrevious;
+        }
+
         return new StatSerializer(stat).perform();
       })
       .then(function (stat) { response.send(stat); })
@@ -50,7 +73,7 @@ module.exports = function (app, model, Implementation, opts) {
       .perform()
       .then(function (result) {
         switch (request.body.type) {
-        case 'Value':
+        case CHART_TYPE_VALUE:
           if (result.length) {
             var resultLine = result[0];
             if (resultLine.value === undefined) {
@@ -63,7 +86,8 @@ module.exports = function (app, model, Implementation, opts) {
             }
           }
           break;
-        case 'Pie':
+        case CHART_TYPE_PIE:
+        case CHART_TYPE_LEADERBOARD:
           if (result.length) {
             result.forEach(function (resultLine) {
               if (resultLine.value === undefined || resultLine.key === undefined) {
@@ -72,7 +96,7 @@ module.exports = function (app, model, Implementation, opts) {
             });
           }
           break;
-        case 'Line':
+        case CHART_TYPE_LINE:
           if (result.length) {
             result.forEach(function (resultLine) {
               if (resultLine.value === undefined || resultLine.key === undefined) {
@@ -89,6 +113,19 @@ module.exports = function (app, model, Implementation, opts) {
               },
             };
           });
+          break;
+        case CHART_TYPE_OBJECTIVE:
+          if (result.length) {
+            let resultLine = result[0];
+            if (resultLine.value === undefined || resultLine.objective === undefined) {
+              throw getErrorQueryColumnsName(resultLine, '\'value\', \'objective\'');
+            } else {
+              result = {
+                objective: resultLine.objective,
+                value: resultLine.value,
+              };
+            }
+          }
           break;
         }